@@ -10,37 +10,35 @@ class _PlatformInterface {
     final lib = ffi.DynamicLibrary.process();
 
     try {
-      final isPresent =
-          lib.lookupFunction<omxpvidpp_is_present_func, OmxpvidppIsPresent>(
-              "omxpvidpp_is_present");
+      final isPresent = lib.lookupFunction<omxpvidpp_is_present_func, OmxpvidppIsPresent>("omxpvidpp_is_present");
       return _PlatformInterface._constructor(isPresent);
-    } on ArgumentError catch (e) {
+    } on ArgumentError {
       return _PlatformInterface._constructor(null);
     }
   }
 
-  final OmxpvidppIsPresent _isPlatformSidePresent;
+  final OmxpvidppIsPresent? _isPlatformSidePresent;
 
   static const _channel = MethodChannel("flutter.io/omxplayerVideoPlayer");
-  static _PlatformInterface _instance;
+  static _PlatformInterface? _instance;
 
   static _PlatformInterface get instance {
     if (_instance == null) {
       _instance = _PlatformInterface._();
     }
 
-    return _instance;
+    return _instance!;
   }
 
   Future<void> init() {
-    return null;
+    return Future.value();
   }
 
   Future<void> dispose(int playerId) {
     return _channel.invokeMethod<void>('dispose', {'playerId': playerId});
   }
 
-  Future<int> create(DataSource dataSource) {
+  Future<int?> create(DataSource dataSource) {
     return _channel.invokeMethod<int>('create', {
       'sourceType': dataSource.sourceType.toString(),
       'uri': dataSource.uri.toString(),
@@ -60,8 +58,7 @@ class _PlatformInterface {
           return VideoEvent(
             eventType: VideoEventType.initialized,
             duration: Duration(milliseconds: map['duration']),
-            size: Size(map['width']?.toDouble() ?? 0.0,
-                map['height']?.toDouble() ?? 0.0),
+            size: Size(map['width']?.toDouble() ?? 0.0, map['height']?.toDouble() ?? 0.0),
           );
         case 'completed':
           return VideoEvent(
@@ -85,8 +82,7 @@ class _PlatformInterface {
   }
 
   Future<void> setLooping(int playerId, bool looping) {
-    return _channel.invokeMethod<void>(
-        'setLooping', {'playerId': playerId, 'looping': looping});
+    return _channel.invokeMethod<void>('setLooping', {'playerId': playerId, 'looping': looping});
   }
 
   Future<void> play(int playerId) {
@@ -102,30 +98,24 @@ class _PlatformInterface {
   }
 
   Future<void> setVolume(int playerId, double volume) {
-    return _channel.invokeMethod<void>(
-        'setVolume', {'playerId': playerId, 'volume': volume});
+    return _channel.invokeMethod<void>('setVolume', {'playerId': playerId, 'volume': volume});
   }
 
   Future<void> seekTo(int playerId, Duration position) {
-    return _channel.invokeMethod<void>(
-        'seekTo', {'playerId': playerId, 'position': position.inMilliseconds});
+    return _channel.invokeMethod<void>('seekTo', {'playerId': playerId, 'position': position.inMilliseconds});
   }
 
   Future<Duration> getPosition(int playerId) async {
-    final result =
-        await _channel.invokeMethod<int>('getPosition', {'playerId': playerId});
-
+    final result = (await _channel.invokeMethod<int>('getPosition', {'playerId': playerId}))!;
     return Duration(milliseconds: result);
   }
 
   Future<void> createPlatformView(int playerId, int platformViewId) {
-    return _channel.invokeMethod<void>('createPlatformView',
-        {'playerId': playerId, 'platformViewId': platformViewId});
+    return _channel.invokeMethod<void>('createPlatformView', {'playerId': playerId, 'platformViewId': platformViewId});
   }
 
   Future<void> disposePlatformView(int playerId, int platformViewId) {
-    return _channel.invokeMethod<void>('disposePlatformView',
-        {'playerId': playerId, 'platformViewId': platformViewId});
+    return _channel.invokeMethod<void>('disposePlatformView', {'playerId': playerId, 'platformViewId': platformViewId});
   }
 
   DurationRange _toDurationRange(dynamic value) {
@@ -138,7 +128,7 @@ class _PlatformInterface {
 
   bool isPlatformSidePresent() {
     if (this._isPlatformSidePresent != null) {
-      return this._isPlatformSidePresent() != 0;
+      return this._isPlatformSidePresent!() != 0;
     } else {
       return false;
     }
